dTemp model normalization issue
The projection models do not need to be re-blazed or adjusted with a scaling factor. The gradient files are in flux per Kelvin. This operation (*b_ratio) messes with the gradient file and will affect the amplitude of the dTemp, specially when there is continuum variations (more or less blue extinction in the optical) from spectrum to spectrum.